Instruction about building and running a C program on RISCV
Follow the instruction to build and run a C program on RISCV.
If it is not building your program check the PATH variable, does it contain path to your riscv-tools installation directory or not? if not then add the path then it will be able to properly locate the cross compiler required to generate the binary of riscv format plus other tools like spike required to run the program etc.
If it is not exporting the environment variable permanently and you have to add the path each time after you restart the system you can add simple shell file (xx.sh) into ect/profile.d/ containing the command to export variable and it will be executed automatically by the system each time when it starts.